Split large file into smaller chunks.
